requirements-dev.txt
aws-cdk-lib>=2.0.0
aws-cdk.aws-ecs-patterns>=2.0.0
constructs>=10.0.0
python-dotenv>=0.19.0
diagrams>=0.23.3
graphviz>=0.20.1
